requirements.txt
pandas
duckdb
mysql-connector-python
pyyaml
streamlit
seaborn
matplotlib

